refactor: change botconfig impplementations
